# docs/openapi/marcus-trading-delta.yaml
openapi: 3.0.3
info:
  title: Marcus Trading API Delta
  version: 0.2.0
  description: >-
    Contract-first delta for Marcus trading integration.
    This version completes High and Medium priority operations and standardizes
    reusable error responses.
  license:
    name: Proprietary
    url: https://marcus.local/license
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [profile]
      operationId: getCurrentUserProfile
      summary: Get current user profile.
      responses:
        '200':
          description: User profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              examples:
                default:
                  value:
                    userId: usr_001
                    username: trader.one
                    email: trader.one@marcus.local
                    role: USER
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/preferences:
    put:
      tags: [profile]
      operationId: updateUserPreferences
      summary: Update current user preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPreferencesRequest'
      responses:
        '200':
          description: Preferences updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/api-keys:
    get:
      tags: [profile]
      operationId: listApiKeys
      summary: List API keys of current user.
      responses:
        '200':
          description: API key list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeySummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [profile]
      operationId: createApiKey
      summary: Create a new API key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/api-keys/{apiKeyId}:
    delete:
      tags: [profile]
      operationId: deleteApiKey
      summary: Revoke an API key.
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '204':
          description: API key revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/login-activities:
    get:
      tags: [profile]
      operationId: listLoginActivities
      summary: Paginated login activity audit logs.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Login activity page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginActivityPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ApiKeyId:
      name: apiKeyId
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: Validation failed.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    InternalServerError:
      description: Internal server error.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OffsetPaginationMeta:
      type: object
      required: [page, size, totalElements, totalPages, hasNext]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
...
    UserProfileResponse:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
    UpdateUserPreferencesRequest:
      type: object
      properties:
        timezone:
          type: string
        locale:
          type: string
        emailNotificationsEnabled:
          type: boolean
    UserPreferencesResponse:
      type: object
      properties:
        timezone:
          type: string
        locale:
          type: string
        emailNotificationsEnabled:
          type: boolean
    ApiKeySummary:
      type: object
      properties:
        apiKeyId:
          type: string
        label:
          type: string
        maskedKey:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
    CreateApiKeyRequest:
      type: object
      required: [label]
      properties:
        label:
          type: string
    CreateApiKeyResponse:
      type: object
      properties:
        apiKeyId:
          type: string
        key:
          type: string
        label:
          type: string
    LoginActivityItem:
      type: object
      properties:
        occurredAt:
          type: string
          format: date-time
        ipAddress:
          type: string
        userAgent:
          type: string
        success:
          type: boolean
    LoginActivityPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoginActivityItem'
        meta:
          $ref: '#/components/schemas/OffsetPaginationMeta'
...
    ErrorResponse:
      type: object
      required: [code, message, status, traceId, path, timestamp]
      properties:
        code:
          type: string
          description: Stable application error code.
        message:
          type: string
        status:
          type: integer
          format: int32
        traceId:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
      example:
        code: BAD_REQUEST
        message: Bot name is required
        status: 400
        traceId: 7dc5e5b8-3136-4b71-8540-b6f85cb3b6b3
        path: /api/v1/bots
        timestamp: '2026-04-10T10:00:00Z'
    FieldValidationError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldValidationError'
      example:
        code: VALIDATION_FAILED
        message: Validation failed
        status: 422
        traceId: 643ad58e-13a8-4f38-9bf1-6be7f0a093f0
        path: /api/v1/paper/orders
        timestamp: '2026-04-10T10:00:00Z'
        errors:
          - field: quantity
            reason: must be greater than 0